backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-pymongo==2.3.0
python-dotenv==0.19.0
flask-login==0.5.0
bcrypt==3.2.0
flask-jwt-extended==4.3.1 